+++ custom_board/arduino/libraries/MCUFRIEND_kbv/MCUFRIEND_kbv.h
@@ -25,7 +25,7 @@ class MCUFRIEND_kbv : public Adafruit_GFX {
 	void     begin(uint16_t ID = 0x9341);                       // you only need the constructor
 	virtual void     drawPixel(int16_t x, int16_t y, uint16_t color);  // and these three
 	void     WriteCmdData(uint16_t cmd, uint16_t dat);                 // ?public methods !!!
-    void     pushCommand(uint16_t cmd, uint32_t * block, int8_t N);
+    void     pushCommand(uint16_t cmd, uint8_t * block, int8_t N);
 	uint16_t color565(uint8_t r, uint8_t g, uint8_t b) { return ((r & 0xF8) << 8) | ((g & 0xFC) << 3) | ((b & 0xF8) >> 3); }
 	uint16_t readID(void);
 	virtual void     fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color);
@@ -36,21 +36,21 @@ class MCUFRIEND_kbv : public Adafruit_GFX {
     virtual void     invertDisplay(bool i);
 
 	uint16_t readReg(uint16_t reg, int8_t index=0);
-	int16_t  readGRAM(int16_t x, int16_t y, uint32_t *block, int16_t w, int16_t h);
+	int16_t  readGRAM(int16_t x, int16_t y, uint16_t *block, int16_t w, int16_t h);
 	uint16_t readPixel(int16_t x, int16_t y) { uint16_t color; readGRAM(x, y, &color, 1, 1); return color; }
 	void     setAddrWindow(int16_t x, int16_t y, int16_t x1, int16_t y1);
-	void     pushColors(uint32_t *block, int16_t n, bool first);
-	void     pushColors(uint32_t *block, int16_t n, bool first);
-	void     pushColors(const uint32_t *block, int16_t n, bool first, bool bigend = false);
+	void     pushColors(uint16_t *block, int16_t n, bool first);
+	void     pushColors(uint8_t *block, int16_t n, bool first);
+	void     pushColors(const uint8_t *block, int16_t n, bool first, bool bigend = false);
     void     vertScroll(int16_t top, int16_t scrollines, int16_t offset);
 
     protected:
 	uint32_t readReg32(uint16_t reg);
 	uint32_t readReg40(uint16_t reg);
-    uint32_t  _lcd_xor, _lcd_capable;
+    uint16_t  _lcd_xor, _lcd_capable;
 
 	private:
-	uint32_t _lcd_ID, _lcd_rev, _lcd_madctl, _lcd_drivOut, _MC, _MP, _MW, _SC, _EC, _SP, _EP;
+	uint16_t _lcd_ID, _lcd_rev, _lcd_madctl, _lcd_drivOut, _MC, _MP, _MW, _SC, _EC, _SP, _EP;
 };
 
 // New color definitions.  thanks to Bodmer
